+++ widgets/tws_widgets/example/lib/core/landing_entries/tws_cascade_section_entry.dart
@@ -3,20 +3,23 @@ CSMPackageLandingEntry _twsCascadeSectionEntry = CSMPackageLandingEntry(
   name: "TWS Cascade Section", 
   description: RichText(
     text: TextSpan(
-      text: "Custom TWS Cascade Section component",
+      text: "Shows a custom main control widget with a colapsable content section.",
     ),
   ), 
   composeLanding: (BuildContext ctx) {
-    return Column(
-      children: <Widget>[
-        TWSCascadeSection(
-          title: "Cascade Example", 
-          mainControl: Container(), 
-          content: Container(
-            height: 300,
-          ),
+    return TWSFLandingFrame(
+      width: 400,
+      child: TWSCascadeSection(
+        title: "Cascade Example", 
+        mainControl: Container(
+          color: Colors.green,
+          child: Text("Main control section"),
+        ), 
+        content: Container(
+          height: 300,
+          color: Colors.red,
         ),
-      ],
+      ),
     );
   }
 );