+++ widgets/tws_widgets/example/lib/core/landing_entries/tws_datetime_picker_entry.dart
@@ -4,15 +4,12 @@ CSMPackageLandingEntry _twsDatetimePickerEntry = CSMPackageLandingEntry(
   name: "TWS Datetime Picker", 
   description: RichText(text: TextSpan(text: "Custom TWS Datetime Picker component"),), 
   composeLanding: (BuildContext ctx) {
-    return Column(
-      children: <Widget>[
-        Center(
-          child: TWSDatepicker<TWSFThemeBase>(
-            firstDate: DateTime(1999), 
-            lastDate: DateTime.now()
-          ),
-        )
-      ],
+    return TWSFLandingFrame(
+      child: TWSDatepicker<TWSFThemeBase>(
+        label: "label example",
+        firstDate: DateTime(1999), 
+        lastDate: DateTime.now()
+      ), 
     );
   }
 );