+++ widgets/tws_widgets/example/lib/core/landing_entries/tws_display_flat_entry.dart
@@ -9,8 +9,10 @@ CSMPackageLandingEntry _twsDisplayFlatEntry = CSMPackageLandingEntry(
     ),
   ), 
   composeLanding: (BuildContext ctx) {
-    return TWSDisplayFlat( 
-      display: "Display example",
+    return TWSFLandingFrame(
+      child: TWSDisplayFlat( 
+        display: "Display example",
+      ), 
     );
   }
 );