+++ widgets/tws_widgets/example/lib/core/landing_entries/tws_dropup_entry.dart
@@ -4,17 +4,15 @@ CSMPackageLandingEntry _twsDropupEntry = CSMPackageLandingEntry(
   name: "TWS Dropup", 
   description: RichText(text: TextSpan(text: "Custom TWS Dropup component"),), 
   composeLanding: (BuildContext ctx) {
-    return Column(
-      children: <Widget>[
-        TWSDropup<int>(
-          item: 2, 
-          items: <int>[1,2,3,4,5,6], 
-          tooltip: "tooltip",
-          onChange:(int item) {
-            print("Selected item: $item");
-          },
-        ),
-      ],
+    return TWSFLandingFrame(
+      child: TWSDropup<int>(
+        item: 2, 
+        items: <int>[1,2,3,4,5,6], 
+        tooltip: "tooltip",
+        onChange:(int item) {
+          print("Selected item: $item");
+        },
+      ),
     );
   }
 );