+++ widgets/tws_widgets/example/lib/core/landing_entries/tws_file_picker_entry.dart
@@ -4,15 +4,14 @@ CSMPackageLandingEntry _twsFilePickerEntry = CSMPackageLandingEntry(
   name: "TWS File Picker", 
   description: RichText(text: TextSpan(text: "Custom TWS File Picker component"),), 
   composeLanding: (BuildContext ctx) {
-    return Column(
-      children: <Widget>[
-        TwsFilePicker(
-          dialogTitle: "File Picker example", 
-          onSelect:(List<XFile> xFiles, List<PlatformFile> files) {
-            print("Selected image: ${files.first.name}");
-          },
-        )
-      ],
+    return TWSFLandingFrame(
+      width: 400,
+      child: TwsFilePicker(
+        dialogTitle: "File Picker example", 
+        onSelect:(List<XFile> xFiles, List<PlatformFile> files) {
+          print("Selected image: ${files.first.name}");
+        },
+      )
     );
   }
 );