+++ widgets/tws_widgets/example/lib/core/landing_entries/tws_frame_decoration_entry.dart
@@ -4,9 +4,11 @@ CSMPackageLandingEntry _twsFrameDecorationEntry = CSMPackageLandingEntry(
   name: "TWS Frame Decoration", 
   description: RichText(text: TextSpan(text: "Custom TWS Frame Decoration component"),), 
   composeLanding: (BuildContext ctx) {
-    return TWSFrameDecoration(
-      child: Container(
-        height: 50,
+    return TWSFLandingFrame(
+      child: TWSFrameDecoration(
+        child: Container(
+          height: 50,
+        ),
       ),
     );
   }