+++ widgets/tws_widgets/example/lib/core/landing_entries/tws_section_entry.dart
@@ -4,9 +4,11 @@ CSMPackageLandingEntry _twsSectionEntry = CSMPackageLandingEntry(
   name: "TWS Section", 
   description: RichText(text: TextSpan(text: "Custom TWS section component"),), 
   composeLanding: (BuildContext ctx) {
-    return TWSSection(
-      title: "Section Example", 
-      content: Container(),
+    return TWSFLandingFrame(
+      child: TWSSection(
+        title: "Section Example", 
+        content: Container(),
+      ),
     );
   }
 );