+++ widgets/tws_widgets/CHANGELOG.md
@@ -4,7 +4,9 @@
     1. Added [Example] dev environment.
     2. Added [TWSFThemeBase] class.
     3. Added the defaults [TWSFDarkTheme] and [TWSFLightTheme] for widgets.
-    4. Added the following TWS widgets: 
+    4. Added Routes system.
+    5. Added [TWSFLandingFrame] wrap for Landing entries.
+    N. Added the following TWS widgets: 
         * [TWSButtonFlat]
         * [TWSSection]
         * [TWSCascadeSection]
@@ -14,11 +16,13 @@
         * [TWSFilePicker]
         * [TWSFrameDecorations]
         * [TWSImageViewer]
-    5. Dependencies added:
+        * [TWSIncrementalList]
+        * [TWSInputText]
+
+    N. Dependencies added:
         * file_picker: ^8.1.7
         * camera_platform_interface: ^2.8.0
         * camera_web: ^0.3.5
-    6. Added Routes system.
         
 - Fixes: